use int64_t instead of long long int

+++ src/scip/lapack_calls.c
@@ -42,7 +42,7 @@
 
 
 /* we use 64 bit integers as the base type */
-typedef long long int LAPACKINTTYPE;
+typedef int64_t LAPACKINTTYPE;
 
 /** transforms a SCIP_Real (that should be integer, but might be off by some numerical error) to an integer by adding 0.5 and rounding down */
 #define SCIP_RealTOINT(x) ((LAPACKINTTYPE) (x + 0.5))
